[fix] 댓글 공감 기능 error 해결

+++ Server/src/main/java/com/e1i5/mentaltal/comment/dto/CommentResponseDto.java
@@ -18,9 +18,9 @@ public class CommentResponseDto {
 
     private LocalDateTime modifiedAt;
 
-    private int viewCount;  // 조회수
+    private long viewCount;  // 조회수
 
-    private int voteCount;  // 공감수
+    private long voteCount;  // 공감수
 
     private long boardId;
