+++ Server/src/main/java/com/e1i5/mentaltal/comment/entity/Comment.java
@@ -35,10 +35,10 @@ public class Comment {
     private LocalDateTime modifiedAt;
 
     @Column(nullable = false) // columnDefinition = "integer default 0"
-    private int viewCount = 0;  // 조회수
+    private long viewCount = 0;  // 조회수
 
     @Column(nullable = false)
-    private int voteCount = 0;  // 공감수 (좋아요)
+    private long voteCount = 0;  // 공감수 (좋아요)
 
     @Transient
     private long bid;
